Extracted level code to separate modules.
